+++ CHANGELOG.md
@@ -7,3 +7,11 @@
 - src-Layout eingeführt
 - YAML-Konfiguration angelegt
 - Grundstruktur für LBM erstellt
+
+## v0.2.0-dev
+
+### Added
+- Health-Check implementiert
+- Prüfung von Python, Restic und Timeshift
+- Prüfung der Passwortdatei
+- Versionsausgabe der installierten Programme